Pagination - plugin first impressions and middleware validations
Details
. First impressions on pagination plugin implementation (TODO: tests);
. Added more validations for pagination middleware "l" and "p" query parameters (also added tests);
. Removed pagination tests from "users" module (moved to pagination module).

TODO
. Implement functional tests for pagination.
