+++ src/shared/pagination/validate.js
@@ -18,14 +18,14 @@ const isInAgreementWithConvention = (string) => {
 
 const constraints = [
   {  field: 'c',  validator: (string) => isJSON(string) },
-  {  field: 'l',  validator: (string) => isNumeric(string) },
-  {  field: 'p',  validator: (string) => isNumeric(string) },
+  {  field: 'l',  validator: (string) => isNumeric(string) && (string >= 1)},
+  {  field: 'p',  validator: (string) => isNumeric(string) && (string >= 1) },
   {  field: 's',  validator: (string) => isInAgreementWithConvention(string) },
 ].map(constraint => ({ ...constraint, ...SHARED_PAGINATION_ERROR_QUERY_PARAMETER_INVALID }));
 
 exports.validate = (queryParameters) =>
   constraints.reduce((err, constraint) => {
-    if (!!err) return err;
+    if (!!err) return err; // Stick with first encountered error.
 
     const { validator, ...error } = constraint;
     const string = queryParameters[error.field];
